// src/client/src/constants/users.ts
export enum UsersActionTypes {
  FETCH_ONE_REQUEST = '@@users/FETCH_ONE_REQUEST',
  FETCH_ONE_SUCCESS = '@@heroes/FETCH_ONE_SUCCESS',
  FETCH_ONE_FAILURE = '@@heroes/FETCH_ONE_FAILURE',

  FETCH_ALL_REQUEST = '@@users/FETCH_ALL_REQUEST',
  FETCH_ALL_SUCCESS = '@@heroes/FETCH_ALL_SUCCESS',
  FETCH_ALL_FAILURE = '@@heroes/FETCH_ALL_FAILURE',

  SAVE_ONE_REQUEST = '@@users/SAVE_ONE_REQUEST',
  SAVE_ONE_SUCCESS = '@@heroes/SAVE_ONE_SUCCESS',
  SAVE_ONE_FAILURE = '@@heroes/SAVE_ONE_FAILURE',

  REMOVE_ONE_REQUEST = '@@users/REMOVE_ONE_REQUEST',
  REMOVE_ONE_SUCCESS = '@@heroes/REMOVE_ONE_SUCCESS',
  REMOVE_ONE_FAILURE = '@@heroes/REMOVE_ONE_FAILURE',
}